@@ -6,17 +6,31 @@
 //  Copyright © 2019 Sean Molenaar. All rights reserved.
 //
 
-import CoreData
 import SwiftUI
+import SwiftData
+
 @main
 struct Fosdem: App {
 
-    // inject into SwiftUI life-cycle via adaptor !!!
-    @UIApplicationDelegateAdaptor(AppDelegate.self) var appDelegate
+    let container: ModelContainer
 
     var body: some Scene {
         WindowGroup {
-            EventListView().environment(\.managedObjectContext, appDelegate.persistentContainer.viewContext)
+            EventListView()
+        }.modelContainer(container)
+    }
+
+    init() {
+        do {
+            container = try ModelContainer(for: Conference.self,
+                                           Room.self,
+                                           Event.self,
+                                           Link.self,
+                                           Person.self,
+                                           Track.self)
+            RemoteScheduleFetcher.context = ModelContext(container)
+        } catch {
+            fatalError("Failed to create ModelContainer for Conference.")
         }
     }
 }
@@ -25,15 +39,12 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
 
     var window: UIWindow?
 
-    static let year = YearHelper().year
+    static let year = SettingsHelper().year
 
     func application(_ application: UIApplication,
                      didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey: Any]?) -> Bool {
         let paths = NSSearchPathForDirectoriesInDomains(.documentDirectory, .userDomainMask, true)
         debugPrint(paths[0])
-
-        DataImporter.context = persistentContainer.newBackgroundContext()
-        RemoteScheduleFetcher.fetchScheduleForYear(YearHelper().year)
 
         return true
     }
@@ -42,49 +53,9 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
 
     func applicationDidEnterBackground(_ application: UIApplication) {}
 
-    func applicationWillEnterForeground(_ application: UIApplication) {
-        RoomStatusFetcher.fetchRoomStatus()
-        RemoteScheduleFetcher.fetchScheduleForYear(YearHelper().year)
-    }
+    func applicationWillEnterForeground(_ application: UIApplication) {}
 
     func applicationDidBecomeActive(_ application: UIApplication) {}
 
-    func applicationWillTerminate(_ application: UIApplication) {
-        self.saveContext()
-    }
-
-    // MARK: - Core Data stack
-
-    lazy var persistentContainer: NSPersistentContainer = {
-        /*
-         The persistent container for the application. This implementation
-         creates and returns a container, having loaded the store for the
-         application to it. This property is optional since there are legitimate
-         error conditions that could cause the creation of the store to fail.
-        */
-        let container = NSPersistentContainer(name: "Fosdem")
-        container.loadPersistentStores(completionHandler: { (_, error) in
-            container.viewContext.mergePolicy = NSMergePolicy(merge: .overwriteMergePolicyType)
-            container.viewContext.automaticallyMergesChangesFromParent = true
-            if let error = error as NSError? {
-                fatalError("Unresolved error \(error), \(error.userInfo)")
-            }
-        })
-        return container
-    }()
-
-    // MARK: - Core Data Saving support
-
-    func saveContext () {
-        let context = persistentContainer.viewContext
-        if context.hasChanges {
-            do {
-                try context.save()
-            } catch {
-                // Replace this implementation with code to handle the error appropriately.
-                let nserror = error as NSError
-                fatalError("Unresolved error \(nserror), \(nserror.userInfo)")
-            }
-        }
-    }
+    func applicationWillTerminate(_ application: UIApplication) {}
 }
